backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.12
ffmpeg-python>=0.2.0
numpy>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
